CHORE: Rename lightyear_ prefixed files — lightyear is now the only network stack [#138]
