+++ crates/client/src/networking/plugin.rs
@@ -4,13 +4,13 @@
 
 use bevy::prelude::*;
 
-/// Networking plugin — all communication now goes through lightyear (LightyearClientPlugin).
+/// Networking plugin — all communication now goes through lightyear (GameClientPlugin).
 /// This plugin is kept as a placeholder for any future non-lightyear networking needs.
 pub struct NetworkingPlugin;
 
 impl Plugin for NetworkingPlugin {
     fn build(&self, _app: &mut App) {
-        // All networking systems are now in LightyearClientPlugin.
+        // All networking systems are now in GameClientPlugin.
         // This plugin is intentionally empty after the tungstenite removal (#138).
     }
 }
