--- crates/server/src/networking/server/lightyear/mod.rs
@@ -1,3 +1,0 @@
-pub mod action_handler;
-pub mod bridge;
-pub mod systems;
