+++ crates/server/src/networking/server/mod.rs
@@ -1,2 +1,4 @@
+pub mod action_handler;
+pub mod bridge;
 pub(crate) mod handlers;
-pub mod lightyear;
+pub mod systems;
