+++ crates/shared/src/protocol/mod.rs
@@ -2,9 +2,9 @@ pub mod bulk_compress;
 pub mod channels;
 pub mod components;
 pub mod compression;
-pub mod lightyear_messages;
-mod messages;
+pub mod messages;
+mod data_types;
 pub mod netcode_config;
 pub mod plugin;
 
-pub use messages::*;
+pub use data_types::*;
